requirements.txt
pandas
selenium
dotenv
deepl
openai
google-cloud-translate
